Create properties file to control logging.
These files are put into the class path, causing Commons Logging to use Log4j instead of SLF4J, and configuring the Log4J appender.
This works only for the Task itself. If the code that initializes the Job (eg. BuildIndex,java) tries to write to the log, it fails because of some incompatibility in SLF4J libraries, I got around this by deleting the SLF4J JARs from the hadoop/lib directory. This shouldn't be necessary, but I haven't found the correct incantation to configure Commons Logging in the main Hadoop process.